fix(tracemetrics): Progressively load metrics timeseries
